Improved Thread.yield / thread preemption (#1533)
* PR#7669 : under Linux, use nanosleep(1) as the implementation of Thread.yield.
This results in a scheduling of threads that is much more fair than before.

* Add a testpreempt.ml test to check that thread preemption is fair.
This is not yet the case for the Windows ports, so disable the test on Windows.

* Use Unix.sleepf as an implementation of Thread.delay instead of Unix.select.
Unix.sleepf can have better resolution than Unix.select.

* Fix signal.ml test
Add "wait" in the runner shell script in order to make sure that the process has properly exited
before its output is compared with the reference output.